+++ CHANGELOG.md
@@ -8,8 +8,11 @@ Tapestry currently operates a one month release cycle between releases, see the 
 
 ## [1.0.12] - Unreleased
 
+### Fixed
+- #272 Disable self-update command if not executed within a phar
+
 ### Added
-- #157 Added lock file support so Tapestry doesn't run concurrently.
+- #157 Added lock file support so Tapestry doesn't run concurrently
 
 ## [1.0.11] - 2017-11-23